+++ packages/client/src/network/http/uesr.ts
@@ -1,1 +1,26 @@
-export function create() {}
+import { CreateUserDTO, LoginUserDTO } from '@hello/common';
+import { post } from './request';
+
+/**
+ *
+ * @param dto
+ * @returns
+ */
+export function create(dto: CreateUserDTO) {
+  return post<CreateUserDTO, {}>({
+    url: '/user/create',
+    data: dto,
+  });
+}
+
+/**
+ *
+ * @param dto
+ * @returns
+ */
+export function login(dto: LoginUserDTO) {
+  return post<LoginUserDTO, {}>({
+    url: '/user/login',
+    data: dto,
+  });
+}