+++ packages/server/src/filters/httpException.filter.ts
@@ -25,7 +25,7 @@ export class HttpExceptionFilter implements ExceptionFilter {
     };
 
     // 设置返回的状态码， 请求头，发送错误信息
-    response.status(status);
+    response.status(200);
     response.header('Content-Type', 'application/json; charset=utf-8');
     response.send(errorResponse);
   }