+++ packages/server/src/modules/user/dto/loginUser.dto.ts
@@ -1,18 +1,13 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsNotEmpty } from 'class-validator';
-import { CreateUserDTO as ICreateUserDTO } from '@hello/common';
+import { LoginUserDTO as ILoginUserDTO } from '@hello/common';
 
-export class CreateUserDTO implements ICreateUserDTO {
+export class LoginUserDTO implements ILoginUserDTO {
   @ApiProperty({ description: '用户账号' })
   @IsNotEmpty({ message: '用户账号必传' })
   account: string;
 
-  @ApiProperty({ description: '用户名称' })
-  name: string;
-
   @ApiProperty({ description: '用户密码' })
+  @IsNotEmpty({ message: '用户密码必传' })
   password: string;
-
-  @ApiProperty({ description: '确认用户密码' })
-  rePassword: string;
 }