+++ packages/server/src/modules/user/user.service.ts
@@ -1,8 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateUserDTO } from './dto/createUser.dto';
 import { UserEntity } from './entities/user.entity';
+import { CustomResponseCodeEnum } from '@hello/common';
+import { LoginUserDTO } from './dto/loginUser.dto';
 
 @Injectable()
 export class UserService {
@@ -11,18 +13,39 @@ export class UserService {
   ) {}
 
   async create(dto: CreateUserDTO) {
-    const { name, password } = dto;
+    const { account, name, password, rePassword } = dto;
 
-    // const user = await this.userRepository.findOne({
-    //   where: {  }
-    // })
+    if (password !== rePassword) {
+      throw new HttpException('两次密码不一致', CustomResponseCodeEnum.fail);
+    }
 
-    console.log(name, password);
+    const user = await this.userRepository.findOne({
+      where: { account: account },
+    });
+
+    if (user) {
+      throw new HttpException('账号重复', CustomResponseCodeEnum.fail);
+    }
 
     const userEntity = new UserEntity();
+    userEntity.account = account;
     userEntity.name = name;
     userEntity.password = password;
 
     return this.userRepository.save(userEntity);
+  }
+
+  async login(dto: LoginUserDTO) {
+    const { account, password } = dto;
+
+    const user = await this.userRepository.findOne({
+      where: { account: account },
+    });
+
+    if (!user || user.password !== password) {
+      throw new HttpException('账号或密码错误', CustomResponseCodeEnum.fail);
+    }
+
+    return {};
   }
 }